requirements.txt.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
--allow-external mysql-connector-python
mysql-connector-python